refactor: made top agent a component
