+++ scrybe_frontend/src/pages/LeaderBoard/Leaderboard.jsx
@@ -9,18 +9,19 @@ import LeaderBoardIcon from "./images/leaderboard-icon.png";
 import AgentReport from "../AgentReport";
 import { useSelector, useDispatch } from "react-redux";
 import { LeaderBoard } from "../../redux/features/agents/service";
+import TopAgents from "./components/TopAgents";
 
-const bgMap = {
-  0: "#E6F0FF",
-  1: "#EDF9F0",
-  2: "#FFFDEB",
-};
+// const bgMap = {
+//   0: "#E6F0FF",
+//   1: "#EDF9F0",
+//   2: "#FFFDEB",
+// };
 
-const colorMap = {
-  0: "#006CFF",
-  1: " #4EC264",
-  2: "#FFEB3B",
-};
+// const colorMap = {
+//   0: "#006CFF",
+//   1: " #4EC264",
+//   2: "#FFEB3B",
+// };
 
 function Leaderboard() {
   const leaderboardData = useSelector((state) => state.agent.leaderboard);
@@ -164,7 +165,7 @@ function Leaderboard() {
                   {searchLeaderboard(leaderboard)?.length > 0 ? (
                     searchLeaderboard(leaderboard)?.map((profile, index) => (
                       <>
-                        <LeaderBoardDisplay
+                        <TopAgents
                           key={profile.agent_id}
                           person={profile}
                           index={index}
@@ -230,43 +231,6 @@ function Leaderboard() {
         </div>
       </NewDesignSideBar>
     </>
-  );
-}
-
-function LeaderBoardDisplay({
-  person,
-  index,
-  handleAgent,
-  agent_id,
-  rank,
-  show,
-}) {
-  return (
-    <div
-      className={styles.Profile1}
-      id={styles.border}
-      style={{ background: bgMap[index] }}
-      onClick={() => handleAgent(agent_id, rank, show)}
-    >
-      <div className={styles.Profile_content}>
-        <div className={styles.Profile_img}>
-          <img src={ProfileName} className="" alt="profile1" />
-        </div>
-        <p>ID: {person.str_agent_id}</p>
-        <h2 className={styles.UppercaseName}>
-          {person.firstname.toUpperCase()} {person.lastname.toUpperCase()}
-        </h2>
-
-        <p>
-          No. of calls taken this {person.weekly || person.monthly}:{" "}
-          {person.total_calls}
-        </p>
-        <h1 style={{ color: colorMap[index] }}>
-          {person.average_score} <span className={styles.small_text}>/10</span>
-        </h1>
-        <p className={styles.Agent_position}> {person.rank}</p>
-      </div>
-    </div>
   );
 }
